refactor valodash details code to separate files for easy maintainance
