+++ src/App.jsx
@@ -12,7 +12,7 @@ import CompetitiveProgramming from "./components/CompetitiveProgramming";
 import Education from "./components/Education";
 import Gallery from "./components/Gallery";
 import Footer from "./components/Footer";
-import ValoDashDetail from "./components/ValoDashDetail";
+import ValoDashDetail from "./components/projects/valodash/ValoDashDetail";
 
 // Scroll handler to reset scroll to top on route change, or scroll to section anchors (hashes) on the home page.
 const ScrollHandler = () => {
